Autowired ImageTransferProperties bean Not Configuration. Solved sonar issue duplicate code in CachingRepository classes
